Sort files by file path.
File path is sorted case insensitive.

Looks like metadata is loaded only for files that the are visible to a
user. Some adjustments had to be made to correctly set the files that
are visible to the user with the sorting.

Note that the number of files that we can load metadata for is hard
locked at 1000. Past 1000, I got an exception that the query was too
long.

Also fixed that files past page 1 won't get metadata loaded when you
click "Show All".