Only users with write access can change copyright.
The existing permission system has an edge case where if a file doesn't
already have Copyright metadata, any user with read access can set a new
copyright on the file.  However, once that permission has been set,
either by the read only user or by the author, it cannot be modified by
the read only user.

To fix this, we make sure not to process files that the user doesn't
have write access on. Note that an iterator was used to loop through
allFiles as it's the only safe way to remove objects from a list while
looping through it. Also made sure allFiles was an ArrayList since the
Arrays.asList() returns an implementation of List that doesn't include
the remove method.

Several warning messages were added with this change.
- If a user selects at least one file that they don't have write access
  to, there will be a warning message on the Edit Copyright page saying
  that they don't have permission to change some or all of the files
  they've selected.
- If the user persists and actually submits a file that they don't have
  permission to change, a message saying that they don't have permission
  to change that file will appear.
- If due to lack of permissions, no files' metadata were changed, then a
  message saying that no changes were processed will appear.
- If the user forgot to select any files to change, a warning will
  appear saying no files selected.