Add file filters to file list page.
Adds 3 filters:
Limit to files that have not been copyright tagged.
Limit to files uploaded by me.
Limit to files where I have manage or write access.

Note that the filters are a bit hacky as they rely on submitting a
hidden form to fake a request to the file listing. The hidden form
contains a duplicate of the visible file filters and also has a list of
the original files selected by the user to edit copyright on.

The list of original files selected by the user enables restoring the
original file list of the user decides to deselect some of the filters.