Update alerts database when metadata updated.
Unfortunately, this breaks the generalized nature of the building block
since this is a UBC specific feature. We want to notify the copyright
alerts building block when a file's copyright has been updated.
Copyright alerts indexes files that doesn't have a copyright metadata
tag, so when people update copyright in the metadata building block, the
change should be reflected instantly.

This is done by passing a list of the files updated to a copyright
alerts building block rest API call which will delete the files from its
index. The rest API call is performed by a Spring client.