Filter to show only files that has been linked.
Allow instructors to limit the files shown to only those that has been
linked into the course.

This turned out to be easier than expected due to the accidental
discovery of the undocumented "linked" metadata property that files
have. Just found it while looking at another issue for the alerts
building block. So we just need to check the "linked" property to
determine if files were linked into the course or not.