Apply button for filters, filter tagged refactor.
Change the filter checklist to use a "Apply" button to start filtering.
Previously, clicking on a filter checkbox would automatically apply it.
As the filtering can be pretty slow, it's probably better to let users
to select multiple filters if they want to and then clicking Apply.

The filter tagged refactor is part of an attempt at speeding it up. It
was previously using Xythos utils to get a file's metadata. This turns
out to be unnecessary as you can retrieve a file's metadata from
CSFile's CSEntryMetadata. Unfortunately, this didn't result in a speed
up, but the code does look a bit cleaner, so going to keep it for now.

The MetadataUtil method of retrieving metadata has the advantage of
being able to deal with multiple data types. I don't want to rewrite
that so just going to keep using that elsewhere for now.