Add cookie session id to index update call.
When calling the Copyright Alerts building block to update indexes on
files, it requires an authenticated Blackboard user session in order to
succeed. Luckily, it looks like we can obtain the user's session id from
BbSession. Setting the Cookie http header with the session id seems to
have let us in as an authenticated user.