Refactor print view, fix for Chrome.
Print view wasn't working in Chrome though it worked fine in Firefox. It
looks like it has something to do with the javascript opening a new
window and writing to it. So I removed that part and replaced it with a
form submit system like the file filters. The print view form has the
same params that the javascript used in its ajax request, it submits to
printView.jsp, the form has _top as the target, which means that the
entire page the user was on gets replaced by the print view instead of
popping open a new window.

The too many files warning message used to appear when there are more
than 100 files. Now that we've had some performance increases, I've
increased it to 1000 files.

Note that the print view has a lower limit of 500 for the too many files
warning. The print view needs to retrieve the properties of every single
file whereas the regular view only needs to retrieve them for just a
single page. So the print view is very much slower than the regular
view.