Remove server side webservice call.
When updating the copyright alerts building block, it was performing an
webservice call to the copyright alerts building block on the server
side. I just remembered previously that this was a concern for server
side performance. I've now moved it into the client side so that the
user's browser is the one that issues the request.

I've also increased the update call's performance by making sure that
only files that haven't previously been tagged will be sent out.

Note that I've forced the the ajax call to be synchronus in order to
ensure that the call is made. I've also assumed that the call will be
successful. If it fails, no error is shown to the user. I couldn't find
an easy way to call BB API's error mechanism from the JSP side. If I do
need to show an error, it won't match in style with the other errors
that this building block throws.