Stores files selected from CC in session.
The list of files the user selected from the content collection for
metadata editing will now be stored in session. Previously, it was
forwarded as a hidden form. The hidden form method turned out to be
problematic since it broke paging if the user selects too many files or
files with long names. Basically, paging works by appending the form
parameters to the "next page" url. So if there are too many files or
files with long names, the url size becomes too large, triggering an
exception.

Also fixed a minor bug that when the user uses "Select All", the warning
that the user selected no files pops up.

Fixes #2