Fix no files when accessing from dropdown menu.
Each file in the content collection has a dropdown menu. One of the menu
option is "Edit Copyright Status". When selecting this option from the
drop down, the file is passed using a different parameter name, namely
"path". Have to add a check for this parameter to add to the files list.

Note, this bug also triggers an exception that occurs when no files are
selected due to it trying to load the files list from session. Since no
session has been saved at this point, the session errors out. Fixed this
by making sure that the session value actually exists before trying to
access it.