Fix "Additional Comments" textbox not saving.
Looks like the textbox field name was changed, maybe with the newer
TinyMCE integration we got in SP13?

In any case, there are many params for a textbox field now, each shares
the same field name but with a different postfix to distinguish between
them. So adding a check for the same parameter but with the "text"
postfix is a quick fix for this.

Not sure if this is the best way to handle this, but I think all the
form generation is black boxed undocumentated API, so no way to know
anyways.