Fix can't edit filenames with special characters.
If a file's name has special character, it is percent encoded when
passed as a GET or POST parameter. Normally, this isn't a problem. But
it turns that the URL links to access the building block has an extra
slash in them, e.g.:

https://connect.ubc.ca/webapps/ubc-metadata-BBLEARN//metadata/list?path=/users/john/Class%201%20Section%20199.pdf

This works fine on the dev vm but it seems like the prod server recently
developed an intolerance for double slashes. The prod server issues a
redirect to get rid of the double slash, resulting in:

https://connect.ubc.ca/webapps/ubc-metadata-BBLEARN/metadata/list?path=/users/john/Class%25201%2520Section%2520199.pdf

Note that in the redirect, the path parameter was *again* percent
encoded. This results in a double percent encoded string, which results
in the file being not found.